# backend/openapi.yaml
openapi: 3.1.0
info:
    title: EduFlow
    version: 0.1.0
    description: |
        EduFlow is a role-based education and workshop management backend API designed for Admin, Institutional Admin, Educator, Student, and Public users.

        The system supports the complete academic lifecycle including authentication, content delivery, assessments, certification, analytics, communication, and AI-powered features.

        #### Core capabilities include:

        - **Secure JWT-based authentication** and role-based access control for admin, institutional admin, educator, and student users.
        - **Institution and workshop administration**, including workshop creation, updates, enrollment tracking, assignments, and lifecycle management.
        - **Centralized content management** for modules, study materials, and learning resources uploaded by educators and accessed by students.
        - **Assessment management** for quizzes, tests, and assignments, including question creation, student attempts, answer submission, and result storage.
        - **Automated evaluation support** for objective assessments to reduce manual grading effort and improve consistency.
        - **Certificate generation** for successful workshop completion and public certificate verification using unique verification codes or identifiers.
        - **Performance tracking and reporting** for student progress, workshop statistics, assessment outcomes, completion rates, and trend analysis.
...

...

...

...

...

...

...

...

...

...

...

...

...
components:
    schemas:
...
        AdminDashboardStats:
            properties:
                active_workshops:
                    title: Active Workshops
                    type: integer
                certificates_issued:
                    title: Certificates Issued
                    type: integer
                total_educators:
                    title: Total Educators
                    type: integer
                total_institutions:
                    title: Total Institutions
                    type: integer
                total_students:
                    title: Total Students
                    type: integer
                total_workshops:
                    title: Total Workshops
                    type: integer
            required:
                - total_institutions
                - total_workshops
                - total_educators
                - total_students
                - certificates_issued
                - active_workshops
            title: AdminDashboardStats
            type: object
...
        EducatorDashboardStats:
            properties:
                active_assessments:
                    title: Active Assessments
                    type: integer
                assigned_workshops:
                    title: Assigned Workshops
                    type: integer
                materials_uploaded:
                    title: Materials Uploaded
                    type: integer
                pending_submissions:
                    title: Pending Submissions
                    type: integer
            required:
                - assigned_workshops
                - materials_uploaded
                - active_assessments
                - pending_submissions
            title: EducatorDashboardStats
            type: object
...
        HTTPValidationError:
            properties:
                detail:
                    items:
                        $ref: "#/components/schemas/ValidationError"
                    title: Detail
                    type: array
            title: HTTPValidationError
            type: object
...
        ParentContactDirectoryResponse:
            properties:
                items:
                    items:
                        $ref: "#/components/schemas/ParentContactItem"
                    title: Items
                    type: array
                total:
                    title: Total
                    type: integer
            required:
                - items
                - total
            title: ParentContactDirectoryResponse
            type: object
        ParentContactItem:
            properties:
                parent_email:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Parent Email
                parent_name:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Parent Name
                student_id:
                    title: Student Id
                    type: string
            required:
                - student_id
            title: ParentContactItem
            type: object
        ParentMessageRequest:
            properties:
                body:
                    title: Body
                    type: string
                student_ids:
                    items:
                        type: string
                    title: Student Ids
                    type: array
                subject:
                    title: Subject
                    type: string
            required:
                - student_ids
                - subject
                - body
            title: ParentMessageRequest
            type: object
        ParentMessageResponse:
            properties:
                accepted:
                    title: Accepted
                    type: integer
                failed:
                    title: Failed
                    type: integer
                message:
                    title: Message
                    type: string
            required:
                - accepted
                - failed
                - message
            title: ParentMessageResponse
            type: object
...
        StudentDashboardStats:
            properties:
                average_score:
                    title: Average Score
                    type: number
                certificates_earned:
                    title: Certificates Earned
                    type: integer
                completed_assessments:
                    title: Completed Assessments
                    type: integer
                enrolled_workshops:
                    title: Enrolled Workshops
                    type: integer
            required:
                - enrolled_workshops
                - completed_assessments
                - average_score
                - certificates_earned
            title: StudentDashboardStats
            type: object
...
        ValidationError:
            properties:
                ctx:
                    title: Context
                    type: object
                input:
                    title: Input
                loc:
                    items:
                        anyOf:
                            - type: string
                            - type: integer
                    title: Location
                    type: array
                msg:
                    title: Message
                    type: string
                type:
                    title: Error Type
                    type: string
            required:
                - loc
                - msg
                - type
            title: ValidationError
            type: object
...
    securitySchemes:
        OAuth2PasswordBearer:
            flows:
                password:
                    scopes: {}
                    tokenUrl: /api/v1/auth/login
            type: oauth2
paths:
...
    /api/v1/communication/parent-contacts:
        get:
            operationId: parent_contact_directory_api_v1_communication_parent_contacts_get
            parameters:
                - in: query
                  name: student_ids
                  required: false
                  schema:
                      default: []
                      items:
                          type: string
                      title: Student Ids
                      type: array
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ParentContactDirectoryResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Fetch parent contact directory entries for students
            tags:
                - communication
    /api/v1/communication/parent-email:
        post:
            operationId: dispatch_parent_email_api_v1_communication_parent_email_post
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ParentMessageRequest"
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ParentMessageResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Dispatch parent communication message and log notifications
            tags:
                - communication
    /api/v1/dashboard/admin:
        get:
            operationId: admin_stats_api_v1_dashboard_admin_get
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AdminDashboardStats"
                    description: Successful Response
            security:
                - OAuth2PasswordBearer: []
            summary: Admin dashboard summary stats (admin only)
            tags:
                - dashboard
    /api/v1/dashboard/educator:
        get:
            operationId: educator_stats_api_v1_dashboard_educator_get
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/EducatorDashboardStats"
                    description: Successful Response
            security:
                - OAuth2PasswordBearer: []
            summary: Educator/institution dashboard summary stats (staff)
            tags:
                - dashboard
    /api/v1/dashboard/student/{student_id}:
        get:
            operationId: student_stats_api_v1_dashboard_student__student_id__get
            parameters:
                - in: path
                  name: student_id
                  required: true
                  schema:
                      title: Student Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StudentDashboardStats"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Student dashboard summary stats
            tags:
                - dashboard
...